+++ src/main/java/chess/view/OutputView.java
@@ -21,5 +21,6 @@ public class OutputView {
             }
             System.out.print(".");
         }
+        System.out.println();
     }
 }
