+++ src/test/java/chess/domain/square/FileTest.java
@@ -1,10 +1,9 @@
-package chess.domain;
+package chess.domain.square;
 
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import chess.domain.square.File;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
