+++ src/test/java/chess/domain/square/RankTest.java
@@ -1,9 +1,8 @@
-package chess.domain;
+package chess.domain.square;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import chess.domain.square.Rank;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
