+++ src/test/java/chess/domain/square/SquareTest.java
@@ -1,10 +1,7 @@
-package chess.domain;
+package chess.domain.square;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
-import chess.domain.square.File;
-import chess.domain.square.Rank;
-import chess.domain.square.Square;
 import org.junit.jupiter.api.Test;
 
 class SquareTest {
